update fc-list to current version on buster
This attemps to update the fc-list file that is
said to be outdated by 3.5 years or more.

Yes, in the new version there are also file pathes
in the output but it is the result of the same
commandline that has been used before:

fc-list :fontformat=TrueType

Compare to 2014 at T79424#862626 and T79424.

The new version is from mw2300, on buster, now in April 2021.

Bug: T79424
Change-Id: I6568c69fc92a7beb36f29ccd71bfa0e5ced77cde