Enable new language button for all logged in users outside test projects
This will enable the feature for users on English Wikipedia who have explicitly
opted into the modern Vector experience, but only because English Wikipedia is
not one of our test wikis.

The feature will stay disabled for both logged in and anonymous users for now.

We can safely use the array notation for this configuration now that the train
1.37.0-wmf.3 has rolled out and has no risk of being rolled back.

Bug: T280526
Change-Id: Ia15f09763f5d888a5a636db02dae429c1bb04b65