Use boolean values for wgVectorLanguageInHeader
It needs to be a boolean on group 1 and 2 wikis.
As the train rolls out it will need to be an array.
Avoid declaring it as an array for now, in case we roll
back. wgVectorLanguageInHeader has backwards compatibility
in 1.37.0-wmf.3 for booleans. in 1.37.0-wmf.4 we will switch
to arrays.

Follow up to Ib4a9c4200b17a79c57fdbfcea2342f58f30e462e

Bug: T280526
Change-Id: I5dbc968e1a404a2ea05bfcf94da7671bd7b9c932