Move $wgCentralAuthTokenCacheType from redis_local to mcrouter
The default behaviour in CentralAuth, which we use at WMF until now,
is for the 1-minute CentralAuth login tokens to be stored in the
CentralAuth session backend (which mainly hosts the 24-hour CA session
blobs), which WMF currently set to redis_local.

For multi-dc, we will have to move CA sessions to a replicated
backend (e.g. Kask), but that's fairly high in overhead for such
short-lived token values.

Keeping these in redis_local isn't an option as we do sometimes need
to write these from the secondary DC, and it's out of scope for the
inital multi-dc launch to expose redis cross-dc (and there's a
simplicy argument for intentionally not supporting that).

Instead, we can write the tokens via mcrouter's TLS-capable cross-dc
route to write the token directly to the primary DC, where it can
then be read from by the subsequent login request. These routes
are already used for propagating WANCache tombstones.

As part of T91820, we will optimise the most common cases by routing
certain HTTP requests directly to the primary DC from the edge.

Bug: T278392
Change-Id: Ia4fefb0eef0ab6b85b8a2a8774ce1a464135aad9