Temporarily shorten $wgParserCacheExpireTime from 30 to 21 days
Due to disk space issues, we're temporarily reducing ParserCache expiry.

After:
- this has been deployed, and;
- the purge script ran at least once, and;
- the pc* cluster was re-compacted to reclaim available space, and;
- the DiscussionTools mitigation has been in effect for 21 days.

.. then we will look at where we stand, and ramp this slowly back up.
Either fully (if space allows) or only in part, with the rest of the
ramp up happening after new parsercache hardware capacity is online.

The number "21 days" is based on previous mitigation in 2018
(T167784, T210992) where were settled on 22 days. The DBAs cut it by
an extra day to 21 days in Ia2c78b60f3961.

Bug: T280605
Change-Id: I2ebe9674fb109fbb82f7933b648a432b17b88e10