Simplify mc.php (2/7): Remove mc-labs.php overrides that match prod
Change mc-labs.php to not be a full copy of mc.php, but instead
only the overrides. This should make it less likely for these to
drift unintentionally, as they seem to have a habit of doing.

For example, each of these were done in Beta after already in prod:

* May 2021: Add missing mcrouter-with-onhost-tier (b02cca3059).
* May 2020: Remove purge channels (daf66a0f7c).
* Feb 2020: Add missing mcrouterAware (f592c193f3).
* Oct 2019: Set missing allow_tcp_nagle_delay (41e4a2b885).
* Jun 2018: Nutracker migration (b32b3611ce).
* Jun 2018: Misc settings (5df4176dca).

Also, while at it, add file documentation for the assumed load order
and inclusion point.

Change-Id: Iea94283c538d54ade03a82411cfe2973684e2c71