Simplify mc.php (3/7): Remove $wgMemCachedTimeout setting
We have two "just in case" settings here for 'persistent' and 'servers'
that seem at least theoretically useful in that they would prevent
contacting a memcached instance with wrong intentions, and would
prevent a presistent connection from dangling and potentially having a
a non-localised impact.

However the timeout setting seems harmless to leave unchanged. At
worst, it would have a localised effect on whatever is wrongly
constructing a cache object, by having the MW default timeout (0.5s)
instead of the one we want for mcrouter (0.25s).

Change-Id: Icc2676c23b808db245e3b727aa774eec4deb5e1a