Simplify mc.php (7/7): Define 'wancache-main-mcrouter' unconditionally
== What

Same as the non-WAN configuration for mcrouter, which is also
unconditional. The wiring for Wikitech wiki is still conditional,
same as before.

== Why

This way, the array can be modified from mc-labs.php without it
appearing like the array may be undefined. In practice, this is fine
because we know there is no Wikitech wiki in Beta Cluster, but it
makes the code harder to reason about.

Change-Id: I13646a5557c9b8cb75f7432c5f7065e72c90efef