Add pool counter for automated search requests
Pool is intended to limit requests identified as external automation.
We've had issues both in the past and currently ongoing with
automation from public clouds overloading the pools and our only
mitigation is to block their traffic outright. This pool gives a place
to direct that traffic where it can continue but in a controlled
manner with less impact to our interactive users.

It's not clear what appropriate sizing of this queue is. For now set
it to 15% of the CirrusSearch-Search pool, which these requests would
have otherwise gone to. Simliarly reduce the size of that pool by 10%.
We may need to revisit, but this seems like a plausible starting
point. Queue size is set to 3x workers, same as other search pools.

Bug: T284479
Change-Id: I3e8da19b2a51e469bf2c2d8367a87a7226ce17b4