Update autonyms in wmgExtraLanguageNames
This makes the language names consistent with the names
in the language-data library, which is used in
the UniversalLanguageSelector extension
and with core Names.php.

For kea, the capital letter is removed. See confirmation
in the Phab task.

sjd and sje are removed, because they are now in core Names.php.

For ota, the name is changed from "Turkish" to "Ottoman", which
is unambigous, and which is already used in language-data.
I consulted about it with the Turkic Wikimedians group,
and they said it's fine.

Bug: T284870
Change-Id: I55f2d36aa3c42dc86ee320a7d388138f43f9d346