Remove $wmgWikibaseClientRepoConceptBaseUri
Unused since change Ia47fe7cf86. The comment, which is still valuable,
is moved to $wmgWBRepoConceptBaseUri.

Note that on Beta Wikidata, $wmgWikibaseClientRepoConceptBaseUri used
HTTP, while $wmgWikibaseEntitySources and $wmgWBRepoConceptBaseUri use
HTTPS. (That said, I’m not sure this part of the configuration is used
in WikibaseClient anyways – I only found uses in WikibaseRepo.)

Bug: T257260
Change-Id: I173bf770c98ff788796ddf729e00b22613d0b158