Merge db-codfw.php and db-eqiad.php into db-production.php
With dbctl, these two files became basically identical except for the
list of ParserCache servers. However, there was still the possibility of
them getting out of sync, which caused an outage for trwikivoyage during
the June 2021 DC switchover as db-codfw.php had a typo that was
previously unnoticed.

ParserCache is now in *Services (since d364074a4f), so the files are
practically equal. For the rare case when a wiki is in a different
section in each datacenter, a ternary can be used with $wmfDatacenter to
handle it. An example of this is provided inline.

noc's db.php will now ensure $wmfDatacenter is set properly in case it's
used by db-production.php.

Bug: T260297
Change-Id: Ibb88165fb8dc18659e4c9db707b3926a5f2b85b5