Uninstall Score on private wikis
There is some theoretical risk of leaking private wiki data via Shellboxes
shared with public wikis. The easiest way to avoid that entirely is to not
allow Score to be used on private wikis. Any <score> tags will now just
be treated as plain wikitext instead of rendering the "Score is disabled"
message.

Bug: T257066
Change-Id: I28d36d6db0f429158606878d3c729852d0f296a8