logos/manage.py: Set user-agent on all requests
With the previous code, all requests to upload.wikimedia.org failed
because of rate limiting in the upload frontend
https://gerrit.wikimedia.org/g/operations/puppet/+/fda621e76a/modules/varnish/templates/upload-frontend.inc.vcl.erb#200

This creates a requests session first, to avoid having to define
the user-agent for every single request separately.

Constructing it directly in download(), as that's the only place
making requests as-of today, but I can also make it global.

Bug: T286797
Change-Id: Id741b35142d50f98fc9e4ef14b30c397c1598e1a