Re-enable commonswiki sister search
This reverts commit 76a2ff29ae84c32620bc794e2c58974f6daab4d1.

Commonswiki was removed from sister search in T264053 as a mitigation
for operational issues. That issue has been mostly root caused and
resolved, we should be able to re-enable this feature without causing
negative impact to the service. We will need to monitor after rollout
to ensure that is the case.

Bug: T277225
Change-Id: I5f00d1d4713fe36ceaa0de01e01f36530b0778bf