noc: Expose primary datacenter on conf/
Parse the YAML file to get the primary DC and expose it for humans on
<https://noc.wikimedia.org/conf/>, right under active MediaWiki versions.
If for some reason that regex fails, it'll fallback to "unknown" rather
than bailing out entirely.

Inspired by Tyler's question on I237eee6bcae07, asking for an easy way
for people to figure out which datacenter is primary.

Change-Id: Ic79f58cd6aedacc2c5d587376d917fe23bf8ddd2