Reduce ten seconds from dispatch max time
With migration to systemd timers, the new one gets triggered if the old
one hasn't finished yet and with this system, we will lose a lot of
capacity if one doesn't get triggered.

This ten seconds gap ensures there is time between finishing of a timer
and starting of the new one.

The whole this is a mess and should be replaced with jobs which is
tracked in T48643

Bug: T288175
Change-Id: I45d2427b1ab01ca6a6ed92fad3a509e73bc3d49a