Introduce $wmfUsingKubernetes to help with migration
Some configuration will need to change depending on whether we're
running on Kubernetes or not (see follow-up). On Kubernetes pods,
the SERVERGROUP variable will be prefixed by "kube-". This value
is controlled by the "php.servergroup" Helm variable, which is
passed onto Apache as a SERVERGROUP environment variable.

See operations/deployment-charts:helmfile.d/services/mwdebug/values.yaml
for an example.

Change-Id: I80526360db3923a1de8794f2334643c720886649