Enable 'DuplicateParse' logging bucket
WikitextContent->fillParserOutput() had some debug message whenever it
encounters redundant parsing.  It was introduced back in 2018 to core
with f503995c51c6 for T205369.

We have a sprint to detect more of this and it seems to be a good
opportunity to start collecting the one passing through that code path.

From testwiki I found 51 hits, so it is most probably low traffic.

Change-Id: I3c91d5134c96261bb18875c47254600cb3db423b