profiler: Use the ProfilerXhprof 'running' option when profiling
The profiler.php code already calls tideways_xhprof_enable() two
lines earlier (which we do to capture the multiversion and wmf-config
code in our profiles, as per T180183). Thus we don't need MediaWiki
core to start the underlying profiler a second time, which is causing
bugs for the "Inline profile" (output=text) mode of WikimediaDebug.

Bug: T247332
Change-Id: I25f4b70b9d4b9cb8f0b3a39c6b4b891c5cc68298