Remove NS_MAIN from wgExtraSignatureNamespaces on most 'special' wikis
In 2014, commit 6c6cbb0b9 added NS_MAIN to wgExtraSignatureNamespaces
on all wikis in the 'special' dblist, stating that they "often have
discussions, including their "village pump" equivalent, in ns 0".

That wasn't entirely true: the 'special' dblist includes a lot of
wikis that don't behave that way, including major projects like
Wikimedia Commons and MediaWiki.org. However, it wasn't a problem,
because that config option was only used to add an extra button on
the editing toolbars. Until…

Recently, we used it in the DiscussionTools extension as a proxy for
guessing whether a namespace should use discussion-related features,
and it turned out that actually, the main namespace on many of these
wikis should not.

After this change, only the following 'special' wikis keep the config:
* incubatorwiki
* labswiki
* labtestwiki
* metawiki
* outreachwiki
* wikimaniawiki
* all old Wikimania wikis
* all private wikis – also, some private wikis gain the config

It is removed from the following 'special' wikis:
(real projects)
* commonswiki
* mediawikiwiki
* sourceswiki
* specieswiki
* testcommonswiki
* testwikidatawiki
* wikidatawiki
(weird wikis)
* apiportalwiki
* donatewiki
* foundationwiki
* loginwiki
* thankyouwiki
* votewiki
(closed wikis)
* advisorywiki
* nostalgiawiki
* qualitywiki
* strategywiki
* usabilitywiki

Bug: T291630
Change-Id: Ic13b3c7ed0ef14895ab06cd75b987f8325ef436c