Enable Parsoid API everywhere
It's convenient to provide access to an uncached API for performance
testing purposes. It would be especially nice to have it on mwdebug* so
that profiling can be generated for specific test cases. I don't think
it's an especially concerning DoS vulnerability -- there are many other
slow APIs that would be more difficult to disable in the context of an
incident.

Change-Id: I1898e8dede4560276ea782c837c36ea1b67ccf64