static.php: correctly report a bad request
When the client sends a request that is missing fundamental parameters,
the server should respond with an HTTP code of 400. While it's nearly
impossible to make such a request to our backends (short of connecting to
the php-fpm daemon and making a fcgi request by hand), we should still
report that the error is on the client's side, not the server side.

Change-Id: Icadab8c8d1781bdaa5d948541c80f45e4c846511