Allow copy-upload (by URL) for Wikisource
Since some Wikisources have different copyright policies to Commons,
and copy-uploading can be useful for importing text scans, it would be
helpful to allow users to import these texts.

This enables copy uploading for Wikisources, as well a defining the most
common import sources: the Internet Archive and the IA Upload tool, as
well a toolforge staging area (T292213).

Note: in order to enabled copy-uploads on a specific wikisource, the
upload_by_url right must be assigned to a user group (for example, on
Commons, this is assigned to the 'user' group). Each Wikisource that
wishes to use copy-uploading should request the right to be mapped to
a user group(s) as a separated config change.

Bug: T293205
Change-Id: I844e80e560dd465bb3044ca3b9bc47740e770ef8