Set dispatchViaJobsAllowedClients to null everywhere
This setting was mainly needed for Wikidata, to roll out the new
dispatch process gradually, for only some client wikis at first. Now
that the Wikidata rollout is completed, we can set it to null
everywhere, to reduce complexity in the config. (Commons has no changes
to dispatch anyways.)

$wmgWikibaseDispatchViaJobsAllowedClients will be removed in a separate
change, to ensure that the config changes are synced in the right order.

Bug: T291828
Change-Id: Ia2b63ada6c6873b646a1d52aa38d70bc753ad041