Repair the size of the logo of Kashmiri Wikipedia
This fix consists of the correction of the sizes of the logos of the
mentioned wiki, upon request in the bug also mentioned.

As this bug was also used to place these logos, it will not be
necessary to quote "T293342" in config.yaml as it should already be
mentioned anyway.

Bug: T293342
Change-Id: I71c272ce795ead5ae37198423f5d99ef0a69fd01