Add framework for setting $wgLocalHTTPProxy
To be cautious, I intend to stage the rollout of this feature by wiki,
controlled by $wmgUseLocalHTTPProxy, which is false everywhere for now.

The proxy is set in ProductionServices under the name 'mwapi', matching
the envoy serviceproxy name. It's set to false for Beta cluster, where
it should have no effect.

Bug: T288848
Change-Id: Id041b16418df9544fe0d740235123ebccecf6321