Wikisource: allow copy-uploads from Commons
It's very common to import from Commons to Wikisources (depending on the
local WS copyright policy), because works may not Public Domain in only
one of the US or their home country.

However, the domain is currently not whitelisted for imports, which is
actually breaking PWB's imagetransfer.py script (T294825). Regardless of
that, it would be better to do these uploads directly anyway.

Bug: T294824
Change-Id: Icfe2a49e6e0abbbd9243d2817aad7a040bbe1ab9