Set up beta test environment for QuickSurveys
Adds a new survey to enwiki beta.

Note that associative keys were added as otherwise the 'default'
and '+enwiki' values are merged incorrectly (e.g. the added survey
would blend into the first, unrelated, default survey).

SiteConfiguration::arrayMerge supports adding to a list, but only
if the subvalues are not also an array.

Bug: T293798
Depends-On: Iae07c56d1067ddc675d72ea0e6ad6c0d1bfb58f9
Change-Id: I7ecd8d45b5b052324ee393ef10ed8cabfb389c5b