Wikisource: enable proofreading change-tagging for all Wikisources
This is an essential precursor to the wider task of using change-tagging
to record page proofreading status.

This has already been enabled on mulWS and Beta WS for some time:

The commit that added this option:

https://gerrit.wikimedia.org/r/c/mediawiki/extensions/ProofreadPage/+/702798

The commit that enabled this at Beta Wikisource:

https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/704574

The commit that enabled this at mulWS:

https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/720983/

Enabling this option will not retroactively tag old revisions. That will
require a maintentance script to be run, which does _not_ need to be
run for this change: it will be run later.

Bug: T289140
Change-Id: I8c2e5cc840f6222416cb86a9ba15a9f69aa07cf6