Expand $wgLocalVirtualHosts, update documentation
This list now contains all SUL wikis that aren't closed so cross-wiki
notifications will work. Ideally it should contain all MediaWikis, but
that's left for another day.

Explain what happens if a non-MediaWiki site gets into this list, and
why at minimum, all SUL wikis need to be in it.

Change-Id: I9f38ad0690211390599a39911ac68a6d1db061d8