Move CirrusSearch more_like traffic to eqiad
This traffic initially moved to the codfw cluster as part of the dc
switchover, in I81ed821c63. It's unclear why it's remained here for
some time, potentionally overlooked. During normal operations all
cirrussearch traffic should run from a single cluster. This is
necessary to ensure we can continue operations with a dc out of
rotation.

Change-Id: I08703b5aa0cb26eb9ff62fd29981bafa4190caa9