wikitech: Remove password clear on block
In early 2019 we had a series of incidents involving Developer accounts.
Our first blunt protection against these attacks was to add a handler
for the 'BlockIpComplete' hook that would lock the backing LDAP account
by removing its password (T218654). This was known to be a blunt
instrument at the time, but it was relatively quick and easy to
implement.

Later we added more effective methods for disabling a Developer account
in the form of additional hook handlers which:

* lock the backing LDAP account using native LDAP features (T168692)
* disable linked Phabricator accounts via API calls
* disable linked Gerrit accounts via API calls
* invalidate existing sessions with wikitech

These protections are more comprehensive than the initial password
change lockout mechanism. They also have the benefit of being reversible.
Reversibility is ideal as it allows a WP:BOLD approach to account
blocking with the knowledge that blocks are easily reversible if found
to be in error or after other issues have been attended to.

Change-Id: I7f4b6d674f3cb1cd2f7869fe7b95747c6dd1e3be