wikitech: use ldap-rw.$SITE for ldap access
Move to the new "ldap-rw" naming scheme (from the old "ldap-labs") and
ensure that it will always use the site-local server. I'd imagine
wikitech will follow the normal DC switchover pattern in the Kubernetes
future, and since the LDAP directories are identical and replicated
between sites we should always use the one that is closer latency-wise.

Bug: T295150
Change-Id: I77e0832b1703ba646ef9321fa4b6fa42359606b9