GrowthExperiments: Disable mobile quality gate
Set growthexperiments-addimage-desktop user preference
(hidden opt-in flag for Add Image desktop mode) default on.
Desktop support is out of beta and should be enabled everywhere
but the preference check was accidentally left in. Enable it
by default (it can only be set manually by testers so for normal
users this means it will be enabled for everyone) which has the
same effect as removing the check but doesn't require a code
change and so can be deployed more quickly.
This is a short-term fix.

Bug: T298122
Bug: T300336
Change-Id: I7dc58bfd2d4895b6efe64718d7d9926efbd2cf75