wmf-config: Use __DIR__ instead of re-using an unintended global
A number of files outside wmf-config (such as multiversion and tests)
build a reference to the wmf-config directory and store that in a
local variable which happens to be named like some of our globals:
"$wmfConfigDir".

Within wmf-config we don't need this since we can simply use __DIR__
directly. In CommonSettings.php that isn't being done yet, which
I'll do in a follow-up.

Before we can do that, this commit fixes the files that have started
using CommonSettings.php's conceptually local variable wmfConfigDir,
and relied on it leaking into there as well, and this actually
works right now.

Bug: T45956
Change-Id: Ia61fea4d0dcf86d51547d3132093a336ab3f2e9f