[Beta Cluster] use require_once instead of include for import.php
The import.php file declares a function, and that function is
unconditionally referenced. As such, it is not optional for this
file to exist (so not appropiate to use "include" instead of require).

Also, it does not return an array or some such, so we mustn't actually
require it here, but merely require_once to ensure by this point.

Using require would mean it could be loaded a second time and that
would fail due to declaring the function twice.

Change-Id: I99aabedb2015fe0a38a1dc3a181536eb1e8b5835