InitialiseSettings: General cleanup
Clean up the page so that each line mentions one and only one Phabricator ticket.
This increases the readability of the page and saves time for those who are looking for a specific change.
This is the first patch in a series of patches to be submitted.
The cleaning up task has been divided into several smaller patches to facilitate reviewing for reviewers and deployers.
This patch covers wgAddGroups from the beginning to the end of letter E (etwiki).
It also sorts projects alphabetically.

Bug: T301647
Change-Id: I16a65f0b6056e5be777bd784a77cee0d0d067d8e