InitialiseSettings: General cleanup, wgAddGroups (R-Z)
Clean up the page so that each line mentions one and only one Phabricator ticket.
This increases the readability of the page and saves time for those who are looking for a specific change.
This is the fourth patch in a series of patches to be submitted.
The cleaning up task has been divided into several smaller patches to facilitate reviewing for reviewers and deployers.
This patch covers wgAddGroups from the letter R (rowiki) to the end of wgAddGroups section.
It also sorts projects alphabetically.

Bug: T301647
Change-Id: I8ea70684caa71b1e4ecda85b907f1b3092a07cff