cirrus: Reduce write isolation to only cloudelastic
Writes to CirrusSearch clusters are typically isolated by running each
cluster in a separate job with a partition per cluster. Lately the
infra has been unable to keep up with the rate of jobs we need to
insert to maintain expected write levels.

By reducing isolation on the eqiad and codfw clusters the updates will
be run in-process where the write request happens, only creating new
jobs for retries.

Bug: T295705
Change-Id: Ib1c3093a39fd6e3488bd79cd6e1d65491b052404