Revert "Set StopForumSpam to enforce on the beta cluster"
This reverts commit dd6dd1e78f1e03fe6c5b0a37603649477b1c7d3b.

Reason for revert: This seems to have inexplicably silently broken SFS on beta, at least from what we're (not) seeing in the logs.  I think we'll need to add some additional logging for the onGetUserPermissionsErrorsExpensive to better debug what's going on here.  Testing a few things in shell.php on beta did not yield any immediate or obvious problems with the extension...

Bug: T304111

Change-Id: I6d89740e91f5b48839275338d919a99ffef3b884