Set MW_USE_CONFIG_SCHEMA constant if file exists.
This is a temporary hack to allow us to enable loading of config
defaults via the new SettingsBuilder class on a per-host basis, by
creating the file /var/run/php/use-config-schema on hosts where we
want to test it.

If the file exists, multiversion/defines.php will define the
MW_USE_CONFIG_SCHEMA constant.
If the constant is defined, Setup.php will load the config schema via
SettingsBuilder, instead of including DefaultSettings.php.

Depends-On: I0bd87e2ddde70ba9fede65b8d607eec01b39aed5
Bug: T304460
Change-Id: I85446563c2723394f1ca727e6f92efd34df5c693