List Kartographer static map exemptions and document+flip default
Invert the default to make it clear what the important setting is,
and what the exemptions are.

List non-Wikipedia projects as false to preserve previous behaviour.

The following are technically enabled by this change which I did not
list as exemptions, on the assumption that they have little to no
mapframe use, and afaik no complex cases where the static-and-click
version doesn't already work fine:

* closed wikis.
* special wikis (loginwiki, votewiki, donatewiki, foundationwiki).
* chapter wikis (nlwikimedia, etc; uawikimedia was listed already).
* Wikimania wikis (wikimania2017wiki, ..).
* Beta Cluster-only wikis (deploymentwiki, en_rtlwiki).

Bug: T291736
Change-Id: I22240af06dd5ef1fbe9c692019191bbfabffad88