Always set MW_USE_CONFIG_SCHEMA.
The use of MW_USE_CONFIG_SCHEMA will soon be redundant. But before we
turn on config schema loading in core by default, we want to test it in
production with an easy way to swicth it off again. That's what this
patch is. Previously, we were testing this on individual host, by
checking if /var/run/php/use_config_schema exists.

Bug: T305176
Change-Id: Id3c39626132c82855352d6f6dc6b6f26adb4bd4d