static.php: Fold "unknown" handling into "nohash"
The only distinction this label made was the metric we dispatch
to Statsd/Graphite. While this was somewhat interesting in theory
in the past when the behaviour was different, it is now literally an
ignored parameter that has no effect, not even in terms of cache
header. Let's get rid of this to simplify the code.

I suspect I was the only one looking at that Grafana panel, and it's
been months since I did even that.

Bug: T302465
Change-Id: I0bdf0b4038d8639858f6cb2ee90747d8c40e6294