static.php: Restore short cache for temporary 'mismatch' response
This hasn't worked correctly for a long time. Until yesterday, this
fell into the nohash bucket which was cached for 24h, which is far
longer then a self-correcting deployment should take.

But, with commit this is extended to 1-year. Fix by using a 1-min
cache instead, matching the kind of shortening we do in ResourceLoader
for similar situations.
<https://www.mediawiki.org/wiki/ResourceLoader/Architecture#Change_propagation>

Bug: T302465
Change-Id: Ic21c18b591c5f00899fdee7d0b7bf6b916b5c42f