profiler: Replace deprecated xhgui-collector and mongofill with local class
The upstream code has been simplified a lot since we first imported
this. Take that over for now. There doesn't appear to be a new
standalone version of the Xhgui_Saver_Pdo so for now I've imported
the relevant bits from XHGui itself and formatted it as a local class
to recognise that this is not something straight from upstream but
something we maintain ourselves.

Bug: T308932
Change-Id: I3882be35572adab7bb077407a4f051a0af05960e