CommonSettings: Remove redundant array_search and missing.php ref
This has not been reachable since the introduction of multiversion,
which naturally has to determine which wiki we're on and to find that
the wiki is known in wikiversions.json in order to include MW core.

Without that, it will fail at that level already, and thus not include
MW core if the wiki is unknown, and thus not initialize MW core,
and thus MW in turn won't load LocalSettings.php > CommonSettings.php,
and thus never reach a situation in CommonSettings.php where the wiki
isn't known.

In addition, there is also already a guard at the top of
CommonSettings.php that ensures this (if !$multiVersion, etc.)

The unreachable nature of this check was first reported in 2011
with T32156. After that, support for 'missing.php' was added to
multiversion. MWMultiVersion::getMediaWiki is called from all entry
points, and takes care of rendering missing.php as appropiate.

Bug: T308932
Change-Id: Ic0e134c61d69ec0d8a063d6f384a59ee7876e094