Separate wmgWikibaseTermboxEnabled and wmgWikibaseSSRTermboxServerUrl
This makes it possible to enable the termbox without having an SSR URL,
which is what we would like on Beta (the extra complication of setting
up the server-side rendering service doesn’t seem worth it there).

(Theoretically, it also lets you set the SSR URL without actually
enabling the new termbox, but that’s not a useful combination.)

Doesn’t actually change anything yet, since currently the termbox is
enabled with an SSR URL everywhere.

Bug: T304328
Change-Id: Id1aba95e0f7241ce202b184895c76943d335250a