Remove misleading "disable" of Special:Mostlinkedcategories
The special page in question is non-expensive and generated on the
fly instead of in a background job, so this line does nothing to
disable it besides confusing people who think the page is disabled
and does need a background job (which really does nothing). See
I5b61d127ad for more details.

Bug: T310456
Change-Id: I4e73d5ab0093b6a2a80293679afc3d29d0f77bf5