phpcs: start to fix SpaceBeforeSingleLineComment.NewLineComment
Switch from repo-wide exclusion to just disabling for the
wmf-config/* directory, where a bunch of file have a significant number
of issues and we may want to decide that the rule should be disabled
there permanently (eg in InitialiseSettings inline comments are used
to note which task is associated with a specific setting). Fix issues
in the other directories.

In the process, update and clarify some comments, eg using phabricator
task ids instead of bugzilla task numbers.

Bug: T171115
Change-Id: Id0979fbfdf67d5626ccaab31373f40ed2c7b8984