phpcs: move SpaceBeforeSingleLineComment.NewLineComment exclusions
Disable within each file and explain why, instead of excluding from
phpcs.xml. Each of the remaining files is one where I believe that
enabling and fixing the rule would result in less readable data and
thus the disabling should be kept, except for a few comments that
I moved.

Bug: T171115
Change-Id: I35adaf3d1cff19ae9071aa09270ffd12bf941c9c