Fix unsupported $wgLogos default configurations
The format of $wgLogos is a little weird. Some elements are strings,
some are nested arrays. Null is supported to unset a logo, but false
is not.

This kind of redoes e9ef1adf from September 2021, which was reverted
in I602db50 a few weeks later without any explanation.

The only hint I could find is a comment in SkinModule.php:

 // @todo: Note the beta cluster and other wikis may be using
 // unsupported configuration where these values are set to false.
 // The boolean check can be removed when this has been addressed.

This sounds like we should fix this. Why wasn't it done?

Bug: T310767
Change-Id: Id2712e2af3d8f33fd4e95e00b90be4ffdf78b6e5