Rename wmgWikibaseUseSSRTermbox to wmgWikibaseTermboxEnabled (1/3)
This global controls two separate Wikibase settings in Wikibase.php,
$wgWBRepoSettings['termboxEnabled'] & $wgWBRepoSettings['ssrServerUrl'].
We want to have a termbox without server-side rendering in Beta, so we
should distenangle these settings. As a first step, rename this boolean
flag to match the setting it controls, “termbox enabled”; and the first
change for that is to introduce the new setting.

Note that we don’t add this variable to InitialiseSettings-labs.php,
since the value would be the same anyways (i.e. the current IS-labs
assignment of wmgWikibaseUseSSRTermbox is redundant today).

(This change was already merged once as change I5e830f0257, but not
deployed, hence this resubmit.)

Bug: T304328
Change-Id: Ice224e23bfaff6df1ddbd1ccc3c4aea276838a83