Rename wmgWikibaseUseSSRTermbox to wmgWikibaseTermboxEnabled (2/3)
Having added wmgWikibaseTermboxEnabled as a replacement for
wmgWikibaseUseSSRTermbox in the previous commit (change I5e830f0257),
use the new variable in Wikibase.php now. Both should have the same
value on all wikis.

(This change was already merged once as change I9602e13b45, but not
deployed, hence this resubmit.)

Bug: T304328
Change-Id: I6376b485dffd1fb9d31f358186b4b5139e9c1547