build: Remove redundant defines.php includes from CI build scripts
This was left-over from a previous refactor, and wouldn't work well
anymore if it was actually needed because it's only doing it for
realm 'labs', whereas the prod InitialiseSettings needs it as well.

Fortunately, it isn't a problem because the callers (buildConfigCache,
and buildDBLists, and tests/ after tests/bootstrap.php) already
include defines.php.

While at it, also fix the seemingly-broken require statement
in buildConfigCache.php. It was doing "../../" instead of "/../" like
its sibling files do. This worked by mistake because the first ".."
are seen as non-existent reference to a sibling of the current dir
e.g. "foo/bar.." and because it also went up one level too far, it
cancelled out that mistake, so "foo/bar/../" ended up working the
same as "foo/bar../../".

Change-Id: I796f38d0f04600c534ac0fcdf0156a37ec0160f6