build: Make config gen signature for prod compatible with test
For my amendment of Ibb73b25a2be40cc to work, I'd like to start
moving toward re-using the same config gen code for prod for standalone
uses as well. For this to work we need to either remove the type
hint or make the classes not have differnet names. I'm doing the
latter here.

Change-Id: I60edfb0f606f11b00d496970f5f921847bdfb10f