Add a new Eventgate stream for revision-score events
The ML team is migrating the ORES platform to the newer one,
running on Kubernetes, called Lift Wing. To test the new infrastructure,
the idea is to add a new EventGate stream called
mediawiki.revision-score-test, following the pre-existing
revision-score schema but tailored for a specific model type
(the ediquality one).

The current workflow for revision-score events is the following:
* ChangeProp listens for revision-change events from Kafka, and once one is available,
  it contacts the ORES api with a special /precache URI.
* ORES' precache endpoint does three things:
  1) It scores the rev-id with multiple model types, all listed in its config.
  2) It saves the results in Redis.
  3) It returns the results to ChangeProp.
* ChangeProp then creates a revision-score event, and it sends it to EventGate
  (code in [1]).

Due to how Lift Wing is structured, the ML team wants to move to a more
granular set of event streams, basically one for each (ORES/revscoring)
model type:
- itemsquality
- editquality (goodfaith + damaging)
- articlequality
- draftquality

This change introduced only a generic "test" one
to avoid confusing users, and if the experiment is successfull
we'll proceed with the other ones too.
The idea is to allow Lift Wing to emit events directly
to EventGate, without the need to use any special state/code
in ChangeProp.

The new events will follow the revision-score schema, that
doesn't need to be changed. Long term we'll probably
migrate to something different, based on the work done
in tasks like T308017.

[1]: https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/services/change-propagation/+/refs/heads/master/sys/ores_updates.js

Bug: T301878
Change-Id: I673c45f7e0d8cbfc136e5e656ee99e38b71e9eef