multiversion: Factor out getTagsForWiki() for re-use
This preps for a follow-up to Ibb73b25a2be40, in which I will
add support for showing the dblists on noc.wikimedia.org/wiki.php
in addition to the MW settings.

It is also prep for T169821, which has as the last step remaining to
improve dblist reader performance. The rough plan I have in mind
is to basically have the dblists artefact (from their YAML source)
be dblists.php instead of dblists txt files. And one of the ways
in which we could optimise that format is to basically have that
artefact keyed by wikiID and point directly to an array of tags,
with getTagsForWiki() just being to `return $requiredLists[$dbname]`.

Bug: T169821
Bug: T308932
Change-Id: Ice5302f791fb1d521fcd993f29daa71e1d82d6bd