cirrus: Disable commonswiki writes to cloudelastic
To restore a new copy of commonswiki_file into cloudelastic we need to
delete the old index. Once that index is deleted mediawiki will start
yelling loudly as many writes will fail. Disable sending commonswiki
writes to cloudelastic to ensure we don't trigger a small mountain of
alerts.

This should be reverted once the index is restored.

Bug: T309648
Change-Id: Ice77556a2819a637a6501b262f9fceb6ee25d61a