Revert "cirrus: Disable commonswiki writes to cloudelastic"
This was expected to hvae a few error messages on deployment but
then clean itself up, but we did not notice there was a 3+ day backlog
of jobs and the jobs in the queue all specify this cluster. Need
to let the writes go through and deal with it in another way.

This reverts commit 12ba14bab8f215408244015b39fa587703f4fe6e.

Change-Id: I85c25637a55fe1c4e277ef3388835133cb97fce4