Don't send debug log from test2wiki to testwiki.log
This should make it possible to use test2wiki for performance testing.

I created test2wiki because testwiki was too special and slow. test2wiki
was meant to be as close as possible to being a regular wiki while still
being usable for testing.

Since then, it has sometimes been used to test cross-wiki features. But
you can still get debug logging on test2wiki by using X-Wikimedia-Debug.
I need a way to send requests to it with debug logging disabled, and
X-Wikimedia-Debug does not provide that feature.

Bug: T279664
Change-Id: Ia726bed42168446cb3c7313cf48ce9eb59f8d787