Configure wbsearchentities profile parameter on Test Wikidata (take 2)
This will make the profile parameter of the wbsearchentities (and
query+wbsearch) API available on Test Wikidata: the API is temporarily
coded to make the parameter only available if this setting has been
configured. Later, we’ll make it available on Wikidata too.

This is a resubmit of change Iee6de25983, which had to be reverted
because the code turned out to have a bug. It should work now.

Bug: T307869
Change-Id: I5a7c53790424dedb9acade5292397562d4d7d3f8
Depends-On: I0f9b688140ffa332dada91c4f4f61e1f2b9be0f6