tests: Fix broken CirrusTest test case
This test passed by as multiple issues cancelled each other out.

1. "wikitech" is not a valid dbname, and so no dblist matches it.

2. "wikitech" is also not matching any valid db suffix, and so
   doesn't get any implicit site family tags like "wikisource" from
   "frwikisource".

3. Without knowledge to the contrary, when asked, $wgConf will
   treat it as a dbname and thus match the InitialiseSettings.php
   entry matching that label, as if it was a dbname instead of tag.

As part of I9d363abd7cfef, this test caused a PHP Notice for
undefined offset, as there is no entry for this wiki in
dblists-index.php as it isn't a valid dbname. The intended test case
is presumably for labswiki, just like the case for 'frwikisource'
is one of the wikis matching the 'wikisource' setting.
s

Change-Id: I26b28eedacf77e2b7a689eb0ba5db8a4381f9ce9